requirements.txt
streamlit
langchain
langchain-openai
openai
requests
pandas
pydantic
python-dotenv
langchain-google-genai